Improve mentoring report autosave and make consistent with exam report
